# app/Http/Controllers/Api/LoginController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use App\Models\Otp;
use App\Models\User;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Hash;
use Illuminate\Support\Str;

class LoginController extends Controller
{


    /**
     * Login Api.
     */

    /**
     * @OA\Post(
     *     path="/api/login",
     *     operationId="login",
     *     tags={"Authentication"},
     *     summary="Authenticate user",
     *     description="Authenticates a user with login credentials",
     *      @OA\Parameter(
     *          name="login",
     *          in="query",
     *          description="login",
     *          required=true,
     *           @OA\Schema(
     *           type="string",
     *           example="admin"
     *           )
     *      ),
     *      @OA\Parameter(
     *          name="password",
     *          in="query",
     *          description="password of users",
     *          required=true,
     *          @OA\Schema(
     *          type="string",
     *          example="123456"
     *          )
     *      ),
     *     @OA\Response(
     *         response=200,
     *         description="Success",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example="true"),
     *             @OA\Property(property="data", type="object", properties={}, example={}, description="Empty object"),
     *             @OA\Property(property="message", type="string", example="Success.")
     *         )
     *     ),
     *     @OA\Response(
     *         response=400,
     *         description="Bad Request",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example="false"),
     *             @OA\Property(property="data", type="object", properties={}, example={}, description="Empty object"),
     *             @OA\Property(property="message", type="string", example="Login field is required.")
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthorized",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example="false"),
     *             @OA\Property(property="data", type="object", properties={}, example={}, description="Empty object"),
     *             @OA\Property(property="message", type="string", example="Login or password incorrect.")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation Error",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example="false"),
     *             @OA\Property(property="data", type="object", properties={}, example={}, description="Empty object"),
     *             @OA\Property(property="message", type="string", example="Validation error message.")
     *         )
     *     )
     * )
     */
    public function login(Request $request)
    {
        try {
            $request->validate([
                'login' => 'required',
                'password' => 'required',
            ],
                [
                    'login' => 'Login kiritilmagan',
                    'password' => 'Password kiritilmagan'
                ]);
            $credentials = $request->only('login', 'password');
            if ($request->filled('login')) {
                if (filter_var($request->login, FILTER_VALIDATE_EMAIL)) {
                    $loginType = 'email';
                } elseif (filter_var($request->login, FILTER_VALIDATE_INT)) {
                    $loginType = 'phone';
                } else {
                    $loginType = 'username';
                }
                $credentials[$loginType] = $credentials['login'];
                unset($credentials['login']);
                $user = User\User::where($loginType, $credentials[$loginType])->first();
                if (!$user || !Hash::check($request->password, $user->password)) {
//                if (!$user) {
                    return response()->json([
                        'success' => false,
                        'data' => new \stdClass(),
                        'message' => 'Login yoki parol xato.',
                    ]);
                }
                $tokenId = Str::uuid();
                $token = $user->createToken($tokenId)->plainTextToken;
                $user->token = $token;

                $user = $user->load([
                    'permissions',
                ]);

                return response()->json([
                    'success' => true,
                    'data' => $user,
                    'message' => 'Success.'
                ]);
            } else {
                return response()->json([
                    'success' => false,
                    'data' => new \stdClass(),
                    'message' => 'Login field is required.',
                ], 400);
            }
        } catch (\Exception $exception) {
            return response()->json([
                'success' => false,
                'data' => new \stdClass(),
                'message' => $exception->getMessage(),
            ]);
        }
    }


    /**
     * @OA\Post(
     *     path="/api/auth/login-otp",
     *     operationId="loginWithOtp",
     *     tags={"Auth"},
     *     summary="Login via Telegram OTP",
     *     description="Foydalanuvchi faqat Telegram orqali olingan OTP kodi bilan tizimga kiradi.",
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"otp"},
     *             @OA\Property(
     *                 property="otp",
     *                 type="string",
     *                 example="57148293",
     *                 description="Telegram bot orqali yuborilgan OTP kod"
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Muvaffaqiyatli login",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true),
     *             @OA\Property(property="message", type="string", example="OTP orqali muvaffaqiyatli login qilindi."),
     *             @OA\Property(
     *                 property="data",
     *                 type="object",
     *                 @OA\Property(property="id", type="integer", example=57),
     *                 @OA\Property(property="name", type="string", example="Islombek Abdurakhmonov"),
     *                 @OA\Property(property="username", type="string", example="islombekdev"),
     *                 @OA\Property(property="token", type="string", example="1|asdasdasdadasd"),
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="OTP kiritilmagan")
     *         )
     *     ),
     *     @OA\Response(
     *         response=400,
     *         description="OTP noto‘g‘ri yoki muddati o‘tgan",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="OTP noto‘g‘ri yoki muddati o‘tgan.")
     *         )
     *     )
     * )
     */
    public function loginWithOtp(Request $request)
    {
        $request->validate([
            'otp' => 'required',
        ], [
            'otp.required' => 'OTP kiritilmagan',
        ]);

        $otpCode = $request->otp;

        $userId = substr($otpCode, 0, -6);

        if (!is_numeric($userId)) {
            return response()->json([
                'success' => false,
                'data' => new \stdClass(),
                'message' => 'OTP formati noto‘g‘ri.',
            ], 400);
        }

        $otp = Otp::where('user_id', $userId)
            ->where('code', $otpCode)
            ->where('expired', false)
            ->where('expired_at', '>', now())
            ->latest()
            ->first();

        if (!$otp) {
            return response()->json([
                'success' => false,
                'data' => new \stdClass(),
                'message' => 'OTP noto‘g‘ri yoki muddati o‘tgan.',
            ], 400);
        }

        $otp->update(['expired' => true]);

        $user = User\User::with('permissions')->find($userId);

        $tokenId = Str::uuid();
        $token = $user->createToken($tokenId)->plainTextToken;
        $user->token = $token;

        return response()->json([
            'success' => true,
            'data' => $user,
            'message' => 'OTP orqali muvaffaqiyatli login qilindi.',
        ]);
    }


}